reworked registration and pipelines
